requirements.txt
python-telegram-bot
requests==2.24.0
beautifulsoup4==4.9.1